move atomics into their own module
